feat: improved logging with logrus
Propagated a logrus.Logger/Entry along for the processor, registry and
providers to use for events that may not require to stop processing but
are useful to know otherwise.

Changed some functions to return errors instead of failing with a logger
trying to centralice the logging or returning of the main failing points
in the main components of the application.

Closes #13